a set of nontrivially interdependent changes, all ultimately motivated by the desire to be sure that our command parsing model is sane and can parse even the hard commands.
1. authentication is now handled by an Authenticator class whose
subclasses handle specific SASL mechanisms. the two IMAP command
handlers Login and Authenticate both use the Authenticator system
now.

2. example Authenticators PLAIN and ANONYMOUS. there's a bug here:
PLAIN is not offered in the capability list, but it's accepted if
the client insists. no big problem - it's not worse than Login,
which we do offer. let's deal with it when we have starttls
written.

3. Command states. a command is now Blocked (has been parsed,
can't be executed due to concurrency restrictions), Executing
(obvious) or Finished (obvious). the concurrency testing in IMAP
isn't implemented; all that's implemented is enough to show that
it works and looks okay together with IMAP::reserve().

4. Input stream reservation. Two IMAP commands, AUTHENTICATE and
IDLE, need to reserve the input stream after successful parse and
read some hard-to-classify data. added an IMAP::reserve() to let
them grab and release the input stream.

5. The Command's Arena is now stored in the Command itself, and
pushed/poped properly by IMAP before/after all Command calls.

6. IDLE implemented. mostly to be sure the parse/command framework
made sense.

7. CAPABILITY looks a little bit better - allows sending the
capability response code, without icky repeated code.

8. changed the noop implementation. the comment is as perfectly
descriptive of the new implementation as it was of the old.

whew.

p4raw-id: //oryx/src/main@129