preliminary search.
in Search, the stuff is parsed, but the search isn't actually
executed. want to get things submitted.

in Command, Set and Mailbox, MSNs are parsed properly. mapping
between MSNs and UIDs works.

in the new Query class, nothing is. The class is a bare skeleton.
That's where I'll be heading now.

in IMAP, there is a workaround to allow "new IMAP" during Command
testing.

p4raw-id: //oryx/src/main@290