the fetch code from melkebalanse, massaged until it compiles.
I want to divide fetch in two: the bits that are body[...] in drag
and the ones that aren't. the latter are handled as bools, so a
client that asks "FETCH FULL FULL FULL" receives only one copy.
the former are handled as proper fetches.

this code is untested and does nothing. I just wanted it in the
depot so I can work on it when I'm ready.

p4raw-id: //oryx/src/main@383