make content-disposition go through the tests correctly.
there's a bit of a question... under what circumstances is
content-disposition superfluous? for now, only for single-part
text messages whose specified disposition is "inline" without
arguments. even that seems a bit of a stretch.

as per rfc 2183, all unknown disposition values are treated as
attachment.

there's always a single empty line between base64 parts and the
boundary. that's because other parts may, perchance, not end with
crlf, and we really need a crlf before the boundary.

p4raw-id: //oryx/src/main@1307