add new Jamsettings ORYXUSER and ORYXGROUP: - mention them in manpages. - add Configuration::compiledIn() for them. - use them as defaults for setreuid()/setregid().
add signal handling in logd; catch HUP and reopen logfile.

be a little more paranoid in Server; check that we have neither
'x' nor 'r' access to the jail directory.

add a chroot mode in server, so logd can chroot to the log
directory instead of to the jail directory.

p4raw-id: //oryx/src/main@2538