Segfault-B-Gone, part #1.
We assumed that any Error/Warning from Postgres was related to a
running query. That isn't necessarily true. For example, when a
backend is shutdown by the postmaster, an "unsolicited" Warning
is issued. So we shouldn't try to use the running query unless
we're sure there is one.

This should go into 0.91.

p4raw-id: //oryx/src/main@2752