A quick sketch of the way forward.
Parsing mail to prepare it for injection and reassembling mail
from the database are complementary operations that were being
performed by the same function (HeaderField::parse()). Now we
have the framework for structured fields to be reassemble()d
instead.

This is how, for example, field wrapping and RFC 2047 encoding
will be implemented. We'll also need it to build AddressField
objects from the contents of address_fields.

This passed tests on the first try. Neat.

p4raw-id: //oryx/src/main@50029