work with squirrelmail.
the problem was that we were too fast for poor squirrelmail. we'd
send BYE, OK and then close the connection, doing it all in the
same TCP packet. squirrelmail would see all that, but parse the
BYE and OK only after it detects the EOF, so it though the close
was improper.

p4raw-id: //oryx/src/main@51054