Fixes the remaining Annotation looping problem.
Store::parseAnnotationName() would create an AnnotationName with
a name and with id == 0, and create an AnnotationNameFetcher to
find its id. But when we got any results, the object was looked
up by the returned id, which failed, and a spurious new object
was created, which Store wouldn't be able to see.

This change makes AnnotationNameFetcher use name lookups too.

p4raw-id: //oryx/src/main@51299