More reorganisation: move (some) related things closer together.
This sort of cleanup, fond as I used to be of it, is clearly not
practical any more. I propose we take a leaf out of Postgres's
book and prefer index scans to seqscans.

Just as we have a "1.0" item that contains a list of things that
are described in their own notes, we could have an "Extensions"
item (or Bugs, ms, msconsole, httpd, etc.).

p4raw-id: //oryx/src/main@51371