if a message can't be injected due to severe syntax errors, wrap it as an application/octet-stream with a descriptive error message and inject that. report it in the 250 line text.
add the message id (1234-12-12) to the Received field so it can be
used to find the message-copy file if necessary, and use it for
the wrapper, too. the application/octet-stream is given that name,
so the typical save dialog will suggest that name.

fix the syntax for quoted localparts - it's a quoted-string, not a
quoted phrase.

p4raw-id: //oryx/src/main@51502