If logd (with security=off) receives SIGHUP, LogServer::reopen does the right thing, but the EventLoop unceremoniously shuts everything down when it notices that select returned EINTR.
(There may be other problems with security=on, but I think it's
just the chroot and altered privileges.)

How to fix this? This change is hardly satisfactory.

It dresses up a global bool handledSignal in EventLoop clothing,
and checks on it along with EINTR. reopen() sets it to true, and
the EventLoop resets it to false if it sees an EINTR.

p4raw-id: //oryx/src/main@51511