when a message cannot be parsed, don't barf. instead, wrap it up like the SMTP/LMTP server does, and inject it like that.
move the wrapping code from the SMTP/LMTP server into Message and
make it a little more capable. it now fetches header fields from
the original rather than from a parsed copy in quite a few cases,
which helps arnt-the-user decide that a message is spam and can be
deleted unread.

this seems to, well, it may work or it may not. the SMTP/LMTP
server works as well as before, and I had no mboxes containing
unparsable mail to test that side of it.

p4raw-id: //oryx/src/main@52055