tell the allocator there won't be any pointers in the allocated memory.
in buffer.cpp, we allocate a lot of memory and use only parts of it, or
even none of it, and don't overwrite the unused memory. if that memory
happened to contain old pointer values, the allocator would see the
pointer values (which the buffer owns, but will not use) and infer that
the objects pointer to have a right to live. that's not too good.

p4raw-id: //oryx/src/main@52071