send just one query at a time to postgres. the database may/will send other queries to other handles.
it might be better to send two queries to each? first one to each, and
if there's anything left to do, send one more to each? but in case we
have one heavy and seven simple queries, and four handles, one of the
simple will be held up unnecessarily. hm.

try it this way.

p4raw-id: //oryx/src/main@52134