don't insist on a mime-version field in order to accept content-type: multipart/*
outlook forwards mail incorrectly: it embeds the forwarded message in a
message/rfc822 object, but it removed some "less important" header
fields from the message being forwarded. makes sense, except that
mime-version is removed, which technically changes the meaning of the
message. the meanings of content-type and content-transfer-encoding are
only defined if mime-version is present.

p4raw-id: //oryx/src/main@52182