start remembering which sessions exist for a given mailbox, and send expunges to all of them.
clearExpunge() only touches this mailbox, of course, otherwise the
imap msn/uid calculation become wrong.

if this is the last session for a given mailbox, don't delete it
when it ends. instead, delete it when a new session is created for
the same mailbox. when constructing a session, copy the msn/uid
map from an old session if possible (should speed up my "select
inbox" by about 3s).

p4raw-id: //oryx/src/main@52292