treat GB2312 specially, to get rid of all the malformed spam:
if a bodypart contains bad GB2312 codepoints, convert it to UTF-8,
substituting U+FFFE for each bad codepoint and similar error, and
report the unicode form to clients later.

convert the body text and subject, but not other header fields
(since it's drastic and we have no justification for converting
any other field so far).

p4raw-id: //oryx/src/main@52313