ignore garbage at the ends of mime fields, provided all the essentials are present.
it would be more correct to flag the junk as errors, but the junk
is a little too common for that.

this leaves charset as a border case. 'content-type: text/plain
charset=iso-8859-5' is not parsed for lack of a semicolon, and if
we then really need the charset... what to do?

p4raw-id: //oryx/src/main@52341