when we dispatch events to a connection whose fd is x, clear the bits for fd x so we won't dispatch again.
the bug was as follows: we dispatch events for fd x. in response,
the connection closed itself and did some other things, one of
which involved creating another connection, which would get
sometimes get x as fd. when we return to the event loop, the event
loop would iterate its way to the newly created connection, see
that its fd is x, and give it the same events as it dispatched to
the old connection. oops.

p4raw-id: //oryx/src/main@52396