there were three identical blocks of code to get rid of duplicate fields. merge them, and make the code a little cleaner too. I think the old code could leave the wrong field if the first field wasn't syntactically valid.
when there are two references fields in a message, delete the
second one. this isn't the only reasonable way to treat the
message. it would also be possible to concatenate the values,
leave the last one, or take the firs part of the first field and
the last part of the last field.

I chose to do this because the implementation is simple, and
because with thunderbird 1.5.0.4/scalix, this code ties the
message to the head of the thread.

p4raw-id: //oryx/src/main@52428