If we want to cooperate with Postgres and ident authentication, we have to make sure we're connecting as something backed by a real Unix user.
- Rename ORYXUSER/ORYXGROUP to AOXUSER/AOXGROUP.
- Change their values from oryx/oryx to aox/aox.
- Get rid of DBUSER and use AOXUSER instead.

Only the installer has some leeway in being able to connect as
PGUSER, because it isn't otherwise changing privileges (as the
Servers do in secure()).

So: we create Unix user aox, and connect to the database as an
unprivileged aox user. The installer running as PGUSER can thus
login as aoxsuper if needed (but it isn't, we've seen).

p4raw-id: //oryx/src/main@52561