When the Postgres server is engaged in WAL recovery, it rejects connections with the FATAL "the database system is starting up" message. Once that's done, it functions normally. On hassath's machine, this sometimes causes aox to not start.
This patch is a simplistic draft of a solution. When it sees the
"is starting up" message, it just sleeps for a second and tries
to connect again. This raises a number of questions:

- Should we limit the number of times we try? The infinite loop
  makes me queasy, but WAL recovery is important, and may take
  some unpredictable amount of time.

- The sleep blocks the EventLoop. I'd prefer to avoid that, but
  since we're in startup anyway, it may not matter much.

- How can I avoid the sleep anyway? Since the connection will be
  closed shortly (by both ends, no less), I can't set a timeout
  on it because the connection will be removed by the EventLoop
  when I return, or a little while later.

Perhaps the thing to do is to wait for the server to close the
connection, and if d->sawStartingup is set (or is < n?) when we
see the Close event, just restart the connection without delay?
Or even with the sleep?

p4raw-id: //oryx/src/main@52640