The reason ?= is not usable for many variables is that Jambase sets defaults.
First, Jam reads -sA=B and sets A to B. Then Jambase comes along
and sets A to a default value unless A has been set using -s. Only
after that is Jamfile/Jamrules/Jamsettings read. For this reason,
Jamsettings never sees an empty BINDIR, it always sees either the
value of -sBINDIR=... or the Jambase default.

p4raw-id: //oryx/src/main@52666