fix a couple of some sneaky load-dependent bugs:
we can't parse a new MSN-using command until we've sent all the
responses for the command. waiting until we've finished executing
it won't do, we have to wait for the session updates too.

change the unblockCommand() logic so calling from session works
reliably.

delay parsing a bit on some cases. basically treat unparsed and
blocked together. unparsed was added later, and looked like a
bolted-on kludge. now it's good.

p4raw-id: //oryx/src/main@53831