Fix the commit logic.
If we get past the while loop, and d->qm is done, then there are
no more deliveryRows coming, and deliveryRow must be zero. It is
thus safe to issue COMMIT and set d->qm to zero, because while
the transaction completes, we won't enter the while loop or try
to dereference d->qm.

This is a bit fiddly, but I don't see an obvious flaw. Perhaps
it's better to just introduce a d->committed flag instead?

p4raw-id: //oryx/src/main@53982