Database::removeHandle() caused enqueued queries to fail without removing them from the queue. When a new handle was created, it would never make it past the (non-Submitted) head of the queue, and nothing could ever touch the database again.
This change allows us to recover from a backend shutdown during
which we issued some queries. (The previous change allows us to
recover if we were idle while the backend was unavailable.)

(Some things that issue queries that fail because no handles are
available will still hang, though, because their callers don't
explicitly check for failure, only for some updates. For example
the session initialiser mechanism is like that.)

p4raw-id: //oryx/src/main@54189