The installer changed the configured value of db-user in order to connect as PGUSER. This caused schema.cpp's well-intentioned attempts to grant privileges to DbUser to assign rights to the wrong user if it was the installer that initiated the upgrade.
To let the installer connect as the PGUSER without manipulating
db-user, we need to change Database::setup(); but because that
reuses Configuration:: variables, and because we don't want to
make the value of PGUSER a configuration variable, we need to
use a new enum; but because the actual PGUSER name and password
must be set by the installer, we need a Database::setup() that
takes (int desired, String user, String pass); but because none
of the other callers care, we retain a (desired=0,User=DbUser)
version for convenience.

The rest of the change updates a few callers to keep up, and
cleans up the installer slightly.

p4raw-id: //oryx/src/main@54797