this does NOT go into 2.06 ;)
map some index and constraint names to more human-friendly error
messages. I tried to test this, but one minute was insufficient to find
a case where the upper layers actually allow this to do anything :(

I think we need to test this by creating a long series of Query objects
and verifying Query::error().

p4raw-id: //oryx/src/main@54842