if List throws a Range exception because something walks off the end of the list, that's bad, but generally it handles while running Query::notify().
at present we respond to it by closing the Database object, which
is inappropriate. the problem is probably connected to a
Connection, but not THAT connection.

this change catches the List exception and logs something to
indicate where the problem is, instead of letting the Database
die.

as of this change, neither the Database nor the appropriate
Connection is closed. instead the Query (and Transaction) are set
to Failed, that's all. the client that provoked the bug will
probably get a timeout error of some kind.

later we may find a way to close the right Connection.

p4raw-id: //oryx/src/main@54916