use UString for header field values internally. in particular Query::bind( const UString & ) is used to copy header fields into the database, so we should never, never, never be fooled by any sender into using sender-supplied utf8. update vast numbers of callers etc.
change Parser822::encodedWords() and friends to return UString.

there is a remaining problem here, namely that sequences of
whitespace in subject fields are now always changed to a single
space. I know how to fix that (well), but it's not tightly
connected with this, so I submit this now.

this probably fixes a problem in sieve, haven't tested. see the
diff.

p4raw-id: //oryx/src/main@54963