when we see an unstructured headerfield containing 8-bit characters, it is possible that the field also contains encoded-words. so split it and then convert each chunk to unicode as seems best.
this removes spaces here and there on the converted chunks and
inserts a single space between each chunk. so what. if someone
wants to have consecutive spaces, that's possible in in a
syntactically valid way. I don't care to try to preserve the
number of consecutive spaces in syntactically invalid fields.

simplify the in-reply-to parsing using String::section().

p4raw-id: //oryx/src/main@54982