implement MULTIAPPEND, but don't announce it, and don't attempt any performance improvement of plain APPEND.
the UIDs are assigned more or less at random. the RFC allows that,
which might possibly surprise its author.

IMO we shouldn't advertise it unless and until we add gofaster bits in
the Injector.

this change is just enough that I can write very simple scripts that
append large numbers of messages, and not have to change them later.

p4raw-id: //oryx/src/main@54989