don't try to eliminate sessions in the session initialiser. if we have to work, we might as well use the data for all sessions.
eliminate one more query in the common case: if we know the new
uidnext/nextmodseq from Mailbox and don't need to lock or update the
db, we don't need to select it from the db.

Change IMAP::idle() to return true when the IMAP server is idle, rather
than when it's in idle mode. use the new semantics in ImapSession to
treat idle mode and no command the same way. but ImapSession also knows
the difference, and sends expunge... so the net result is that flag
updates are delayed until the end of idle (and often skipped because an
expunge has removed the MSN for the updated message).

create just a single SessionInitialiser when a Mailbox needs to create
one, instead of 1-n, where n is the number of sessions. the old code
would have several competing session initialisers, one of which would
grab the necessary locks. the others would let themselves be GCed.

eliminateGoodSessions() is dead code now, and possibly/probably broken
anyway. I think what we want is to test whether all sessions can send
their updates without SI work, and if that is the case, we want to skip
ahead in the state machint to where we call setSessionInitialiser( 0 )
and emitResponses().

p4raw-id: //oryx/src/main@55030