ImapSession used to unblock IMAP whenever it tried to emit responses. that wasn't so fortunate. if IMAP::runCommands() somehow told ImapSession to _try_ to emit responses, runCommands() would loop.
after this change, ImapSession instead tries to look for the Command
that would emit the EXISTS/EXPUNGED responses, and if there is one, it
prods that Command to emit its responses. if the Command succeeds in
eitting the necessary responses, it change its state, which in turn
causes IMAP to be unblocked, so the next Command is processed.

IMAP;;runCommand() now only loops if a Command climbs the
Command::State latter.

Since Command::emitResponses() and ImapSession::emitResponses() call
each other, the code also adds a new recursion guard.

p4raw-id: //oryx/src/main@55078