tech imap fetch to trickle responses.
when the fetcher has fetched messages from the database, we want to send
some ASAP, but we don't want to send ALL quickly. instead, we want to make
sure there's a steady stream of responses to the client.

this change tries to ensure that there are few gaps and that the speed
generally increases.

the code can be improved, but I wanted to run it, not make it perfect.