Allow clients to unsubscribe from deleted mailboxes again.
Sebastian Klarmann reported that if he renamed a (\HasChildren) folder
with Thunderbird, it could not see the newly-renamed folder afterwards.
Thunderbird issues the following commands:

    x RENAME a b
    y UNSUBSCRIBE a
    z SUBSCRIBE b
    [...plus UNSUBSCRIBE/SUBSCRIBE for each child mailbox...]

This didn't work, because Subscribe::parse() used Command::mailbox() to
obtain the mailbox, and that function would not allow deleted mailboxes
through. We need to use mailboxName() instead.