Change SaslConnection::close() to log to the connections table.
This doesn't compile, because the servers that include "server" (i.e.
Connection) now depend on "user", and thus on a bunch of other things
too, not all of which can be sanely included everywhere.

So long as the object just contains a pointer to a User, everything is
fine, but we can't call id() without linking in user.o too.