Serve /webmail/archiveopteryx/jquery.js ourselves.
No longer painful, given StaticBlob and some Jam .inc magic.

It's unfortunate that link.cpp grows another static function with a
wart in the form of #include "jquery.inc", but it does work.

(We serve jquery.js now, but FrontMatter::jQuery() doesn't link to it
yet. Is this something the user should ever be able to configure?)

The exact same technique would work for favicon.ico, except that it
needs to be under /, because that's where browsers will look for it.