Add a (stub) EventLoop::shutdownSSL() function.
This doesn't work, because Connection::Listens is set by the Listener,
while Connection::StartsSSL is set by the Listened-for subclass, and
the Listener doesn't have access to an object of that subclass when
it is being created.

In fact, the property can be set only in the subclass, but it really
belongs to the Listener... which knows it only based on the subclass.

Impedance mismatch.