run the spool only every two hours (minimum).
if there's any mail that needs to be delivered sooner, then it will be.
but if we have absolutely nothing to do, we can wait two hours before
doing it.

the exact period doesn't matter, so long as it's longer than the time I
take to ponder my logfile. I have it when I stare at the logfile of an
otherwise idle server and this query arrives. ten minutes is too short,
two hours long enough.