clean up Session updates and the IMAP state machine.
eliminates double flag responses, simplifies the interface between Session
and ImapSession(), simplifies IMAP::runCommands().

Store no longer sends its own responses. Instead it trusts the ImapSession
to send the right flag updates.

Flag updates are now sent before the command that caused them, instead of
between that command's untagged and tagged responses. (If the cause isn't
in this session, then the very next command is picked as cause. It may not
be fair but it's simple.)

The IMAP engine will not retire commands out of order any more. Commands
may be executed concurrently, but their responses will be sent in-order.