Don't keep re-evaluating the sieve while processing RCPT TOs.
If a client issued a hundred RCPT TO:s, a hundred SmtpRcptTo objects
would be created, would each call Sieve::addRecipient(), and all but
one would go back to sleep because another command came first; that
command calls Sieve::evaluate(), which would see hundred recipients
and complete execution with a fair bit of logging. The next object
would call evaluate again, and everything would be repeated.

This change just skips the call to evaluate() if the Sieve is done().