Skip another call to Sieve::evaluate() if possible.
It is my understanding that if Sieve::done() returns true, the addition
of more data (e.g. the setMessage() call in SmtpBody) will not alter the
decision already arrived at. If any of the scripts under consideration
could depend on the message data, done() will return false.

(Unfortunately, that means that the earlier patch doesn't go far enough.
If one of the hundred recipients *does* have a message-dependent sieve,
the O(n²) logging will recur.)

Arnt?