run the spool as soon as the db has completed its startup work
I think we can do the same for shutdown... when we shut down, close the
listeners, ask the db to poke a helper when it becomes idle, and have the
helper shut down. as soon as everyone has finished their transactions
we'll shut down. aox restart can start the new server as soon as the
listeners are gone.