try to reduce postgres server churn.
this has two parts. first, we avoid closing the connections unless we
really have too many. if we've needed ten connections in the past
db-handle-interval, we'll keep ten connections open.

second, when we do close one, make it the youngest connection instead of
the oldest one. this means that we'll tend to close connections we've
never used, and keep connections that we have used (and that have prepared
statements).