adjust database handle closing
now that we use listen/notify we need to reduce databse handle churn, so
that's what we try to do. the number of handles will stay high for one
db-handle-interval, then during the second interval after a peak it can
drop to the new level. we generally try to use the oldest connection(s)
and let the newest connection(s) be closed.