unfuck rename a little bit
it's fucked still, but less. now there's only a small race condition:
rename expects listen/notify to be done _quickly_ and set up the mailbox
tree correctly _quickly_. that's mostly the case. but if it's not and an
imap client pipelines beyond the boundaries in the rfc, shit can happen.