Release savepoints after we're done with them.
If we're going to inject a lot of messages, we'll create a potentially
very large number of savepoints. We can make things easier on Postgres
by releasing them when we don't need them any more.

In theory, we could just keep reusing "savepoint a", either releasing
it or rolling back to it, but I kept the numbers for debugging.

Releasing the bodyparts savepoints makes the most difference (because
there are so many of them), but we do this for the New*Creators too,
because even though there will be only one savepoint to release, the
transaction may go on for a long time afterwards.