Convert MidFetcher/selectMessageId/insertMessages().
insertMessages() is easy, but it needed m->databaseId(), which is set by
the selectMessageId() and the MidFetcher, which needed to be taught how
to take a list of messages and set the id on each one.

It's a pity that we can't use a COPY in selectMessageId(), because then
we would have no way to fetch the ids of the just-inserted-rows without
locking and stuff. I decided not to worry about it for now, because the
common case is 1 message, and the larger referring tables all use COPY.