Overhaul selectUids()/UidFetcher.
This is really the core of the new multi-message Injector.

Instead of claiming one uid and one modseq in each of n mailboxes, we
now claim m uids and one modseq in each of n mailboxes, where m is the
number of messages to be injected into each mailbox in turn.