take two seconds to shut down.
the shutdown period should probably be configurable.

not sure what to do about logd. if the old logd closes its listener so
that we can start a new one, what happens to the file when both write?
perhaps the best way out of that is to use syslog by default and let it
handle that.

aox stop/restart can connect to the imap port, wait for * BYE, and when
that happens the IMAP port will usually be free.

I haven't tested this. I'll test it ad-hoc before I push, but I don't know
how to test it systematically.