Get rid of the Format argument to Query::bind().
Mostly. bind( uint, String, Format ) has been retained because the
Injector uses it to bind text/data when inserting into bodyparts.

COPY is either all-text or all-binary; and apart from the exception
noted above, we use binary format only when doing COPY, so it's safe
to set the format based on the "with binary" part of the query string.

(udoc now complains about the definition of Query::Format.)