Get rid of Flag::create() and replace it with a real class. Make it work.
The hacky Query * return value is now replaced by the Transaction that
matters: If the FlagCreator wedges the transaction and can't recover
(which could happen because q was used for several purposes and they could
be mixed up), then it simply lets its owner take care of the failed
transaction in the usual manner.

AnnotationNameCreator needs the same treatment, but I'd like you (crab) to
do it, so the state machine gets another careful look.