try to rework insertMessages() into using a temporary table.
slightly less than an hour's work, and it seems to work, and it seems to
be pleasantly fast. now we can drop the 'unique' bit on bodyparts and
james mansion will be happy.

(abhijit: feel free to drop this if it goes in the wrong direction, or use
it, as you please.)