make a new superclass to do everything that's common to FlagCreator, AnnotationNameCreator and FieldNameCreator.
adapt FlagCreator to use the new class.

three more changes to come: adapt AnnotationNameCreator. adapt
FieldNameCreator. use a List<HelperRowCreator> instead of individual
pointers in Injector.

Injector would be happy if AddressCreator were a subclass of
HelperRowCreator... if that can be done. I too would be happy. those
addresses lookups are the last long queries we have, and the only log
lines are the only ones that syslog truncate.