Don't send FLAGS updates
The RFC asks for FLAGS responses at SELECT/EXAMINE time. We send that.  It
doesn't ask for more, and we don't send more now, even though mrc and Timo
think we ought to, because it was really, really expensive. Two thirds of
opening my mailbox went towards sending FLAGS responses.