make sure Bodypart::header() returns what it ought to.
If a Message has a single child and isn't a multipart, the the Message's
Header also acts as the child's Header. This is the case that corresponds
to a simple text/plain message: no separate mime header.

this change makes that happen, in a not terribly intuitive way. when we
read the header fields from the db, and can infer that the message has a
single child, then we make sure the child exists, and when we look up that
child, it has no header and it ought to have one, we link it to the
Message's header.

I'm sure this could be done cleaner. Perhaps we could use something like
  Header * Bodypart::actingHeader() const
which returns its own or the surrounding Message's.