Update our refuse-reject implementation to match that which becomes RFC.
Really simple, but it doesn't make MDNs. I'm not sure how to decide
whether to issue an MDN or a DSN. The information is avaiable at the
start, but not at the end of Sieve processing, and the MDN/DSN has to
be made at the end...

This may or may not go into 3.0.0.