Add a new db-schema configuration variable.
1. DBSCHEMA in Jamsettings, empty by default.
2. Set DBSCHEMA in EngraveConfig and two Jamfiles.
3. (Terse) Documentation for the new config variable.
4. The installer takes the default value from Jamsettings, allows it to
   be overriden from the command line (-S); and if no value was given,
   it uses the value from the configuration file, if present.

I thought this was unnecessary, but changed my mind after much thought.

For one thing, "aox grant privileges" and the like need to know if they
are dealing with objects in a particular schema, but if we just set the
search_path and forget about it, we can't know. This is a problem also
because objects in other schemata always show up in the catalog tables.

Besides that, however, changing the search_path was easy, but ultimately
not sensible behaviour. Something that requires us to put dire warnings
in the manpages is suspicious at best.

This will require follow-up changes to set the search_path for each
connection instead of by default. But, once done, that will be much
more predictable and sensible: It makes _sense_ that if you change
the db-schema value, you are using a different schema; and you can
change it back to undo the change. And you always know which schema
is going to be used when the server is started.