Set the search_path per-connection, not per-user.
Since we now have a record of which schema we're installed into, we can
set the search_path as a connection parameter. This removes the need for
a scary "you may hide all your database objects" warning that would need
some surgery to correct: if you hide your objects, just reset the change
to db-schema, and you have them back again.

This is safe to do because, as far as I can tell, specifying a schema
that does not exist in your search_path does not provoke an error when
you do it as a connection parameter. Otherwise we would have to jump
through some hoops to avoid specifying it in the installer (where we
may need to create the schema in the first place).