Avoid writing plaintext LOGIN passwords to the log file.
Done this way on the basis that it fixes the obvious case (a login x y),
and that login is an unlikely source of problems that we'd desperately
need the first line to debug anyway. Suppressing the line seems better
than trying to find and obscure the password (cf. "a login x {7}").

Doesn't do anything about AUTHENTICATE initial-response etc.