smtpclient could work itself into a no-good state. fix.
there was a race. if the smarthost closed the connection JUST when a new
message was spooled for deliveries, then smtpclient would be asked to
deliver, find itself unable to, reconnect, but never actually get anything
done, there would just be a sequence of five-minute connections doing
nothing.

this fixes the race by simplifying the notification model. the owner and
user (spoolmanager and deliveryagent) are notified (newly via notify()
instead of execute() whenever the smtpclient changes state in react, no
more, no less.