make the fetcher use a transaction if it is to fetch dynamic data, to be sure that the reported modseq is in sync with the other queries. (I'm not sure this is sufficient, but I also don't care greatly.)
allow setters to provide a transaction, of which the fetcher will use a
subtransaction (so the fetcher gets exclusive use of the transaction for a
little while).

use the new setter for flag-update fetches, so noone will be able to sneak
in and expunge the message after fetch has found it, but before the
fetcher has retrieved data.