make the AddressCreator generally more similar to HRC.
important difference: the AC has an optimisation where it skips using a
savepoint for the first select. because of this, the caller has to look at
whether the AC is done, not just send more queries (or? hm? not sure.
maybe it would work anyway, due to the magic of subtransactions.) and the
AC may have to notify the caller itself, not rely on sub->commit().