continue splitting messages and mailbox_messages.
the trivia fetcher conflated the two, which led to strange failures and
race conditions, all very unusual but decidedly buggy.

as a bonus change, the injector now records that the message it caches
contains information. this makes whip happy, not sure why. but from now on
the cache may actually lead to speedups.