Rewrite MessageSet to improve some Os, particularly index() and value().
Remove MessageSet::where() and update callers.

Avoid calling MessageSet::add( 1, UINT_MAX ) since that now consumes lots
of RAM.

The new MessageSet uses a number of blocks, each containing one bit for
each number in the set. The memory usage is proportional to the number of
UIDs in the set, but tends to be low if the UIDs cluster together.

index() and value should be faster than they were, but can still use some
optimisation. but this works, so I submit it.