if we're using pg 8.1, we probably benefir from using enormous batches
pg misplans some of our queries (the header_fields one in particular) and
uses seqscans too much. if we're going to seqscan, we might as well pick
up everything... except that that may cost us too much RAM. so we try to
be clever and set the batch size high only for non-body fetches.