make refreshMailboxes work at once in this server AND NOTIFY OTHERS.
change most notifies to call refreshMailboxes, so they'll learn at once
and also tell othjers.

after this change, it's important to call refreshMailboxes() only at
commit time, since if an error occurs after rM() and before commit, then
the in-memory tree would get out of sync with the on-disk tree.

the two notifies in Mailbox are notifies because the transaction may break
after them.