use subtransactions to update the mailbox tree, sessions, and flags.
when we update the databases, we generally need to do some querying and
reporting. this change makes those go into the updating transaction, so
that the updates happen at completely predictable times.

seems to work, although the number of savepoints may be too high. some
state machine may be bad.

there is a problem - if e.g. the injector's transaction fails to commit,
then we've already updated sessions. the only reasonable way out of that
is to abort those sessions.