change lock order so that if we lock both mm and mailboxes, we lock mm LAST.
a non-peek fetch may now lock the mailbox, if it retrieves both the modseq
and needs to set \seen. unfortunate, but if we didn't do this, then the
store could lock the mailbox after fetch locks mm, and the sub-fetch used
to report blah could deadlock against this or another fetch. (even against
itself because the store wasn't even a subtransaction until now.)

store is a subtransaction so that a modseq non-peek fetch won't deadlock
against itself via store and the flag update.

store and expunge lock mailboxes before mailbox_messages.